Fix the x_ddot feature
I focues on only imu for state estimation which is not an easy task.
used the constant acceleration model for estimating velocity which is not accurate because the robot retains its speed even
when the robot stops so i had to utizlied use_control (cmd_vel) topic to overwrite the linear x speed if its 0
the main reason is because when the robot is moving with constant velocity or when its resting the accelerationX is zero so i had to distinguish between the two
if i don't do this the robot moves even when its resting just because of the noise in acceleration X of the imu. so it just sets the velocity and the estimation moves.